### 2025-06-18 #8428 @editorialbot
[PRE REVIEW]: Oetqf: A Python package for quasi-dyanmic earthquake cycle simulation
**Submitting author:** @shipengcheng1230 (<a href="http://orcid.org/0000-0003-3558-458X">Pengcheng Shi</a>)
**Repository:** https://github.com/shipengcheng1230/Oetqf.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.3.0
