### 2025-06-19 #8428 @shipengcheng1230
Hi @kthyng, thank you for the quick feedback. It seems there are some doi issues with the references which I will double check later. Meanwhile, I'd like to suggest the following non-conflicted reviewers (in the field of geophysics):
- `[@krischer](https://github.com/krischer)`
- `[@calum-chamberlain](https://github.com/calum-chamberlain)`
- `[@ziyixi](https://github.com/ziyixi)`
- `[@niyiyu](https://github.com/niyiyu)`
- `[@akinremisa](https://github.com/akinremisa)`

Last, the main repository for review is https://github.com/shipengcheng1230/Oetqf.jl. At the same time, we also have two companion packages supporting this, which are https://github.com/shipengcheng1230/GmshTools.jl/ and https://github.com/shipengcheng1230/GeoGreensFunctions.jl.

Please let us know if there is anything else needed!
