### 2025-07-04 aed13c6
update doi of all references
