+++ paper/paper.bib
@@ -1,55 +1,74 @@
 @article{allison2018earthquake,
   title     = {Earthquake cycle simulations with rate-and-state friction and power-law viscoelasticity},
-  author    = {Allison, Kali L and Dunham, Eric M},
-  journal   = {Tectonophysics},
   volume    = {733},
-  pages     = {232--256},
+  issn      = {0040-1951},
+  url       = {http://dx.doi.org/10.1016/j.tecto.2017.10.021},
+  doi       = {10.1016/j.tecto.2017.10.021},
+  journal   = {Tectonophysics},
+  publisher = {Elsevier BV},
+  author    = {Allison, Kali L. and Dunham, Eric M.},
   year      = {2018},
-  publisher = {Elsevier}
+  month     = may,
+  pages     = {232–256}
 }
 
 @article{barbot2017displacement,
-  title     = {Displacement and stress associated with distributed anelastic deformation in a half-space},
-  author    = {Barbot, Sylvain and Moore, James DP and Lambert, Val{\`e}re},
-  journal   = {Bulletin of the Seismological Society of America},
+  title     = {Displacement and Stress Associated with Distributed Anelastic Deformation in a Half‐Space},
   volume    = {107},
+  issn      = {1943-3573},
+  url       = {http://dx.doi.org/10.1785/0120160237},
+  doi       = {10.1785/0120160237},
   number    = {2},
-  pages     = {821--855},
+  journal   = {Bulletin of the Seismological Society of America},
+  publisher = {Seismological Society of America (SSA)},
+  author    = {Barbot, Sylvain and Moore, James D. P. and Lambert, Valère},
   year      = {2017},
-  publisher = {Seismological Society of America}
+  month     = mar,
+  pages     = {821–855}
 }
 
 @article{barbot2018asthenosphere,
-  title     = {Asthenosphere flow modulated by megathrust earthquake cycles},
-  author    = {Barbot, Sylvain},
-  journal   = {Geophysical Research Letters},
+  title     = {Asthenosphere Flow Modulated by Megathrust Earthquake Cycles},
   volume    = {45},
+  issn      = {1944-8007},
+  url       = {http://dx.doi.org/10.1029/2018gl078197},
+  doi       = {10.1029/2018gl078197},
   number    = {12},
-  pages     = {6018--6031},
+  journal   = {Geophysical Research Letters},
+  publisher = {American Geophysical Union (AGU)},
+  author    = {Barbot, Sylvain},
   year      = {2018},
-  publisher = {Wiley Online Library}
+  month     = jun,
+  pages     = {6018–6031}
 }
 
 @article{barbot2018deformation,
-  title     = {Deformation of a half-space from anelastic strain confined in a tetrahedral volume},
-  author    = {Barbot, Sylvain},
-  journal   = {Bulletin of the Seismological Society of America},
+  title     = {Deformation of a Half‐Space from Anelastic Strain Confined in a Tetrahedral Volume},
   volume    = {108},
+  issn      = {1943-3573},
+  url       = {http://dx.doi.org/10.1785/0120180058},
+  doi       = {10.1785/0120180058},
   number    = {5A},
-  pages     = {2687--2712},
+  journal   = {Bulletin of the Seismological Society of America},
+  publisher = {Seismological Society of America (SSA)},
+  author    = {Barbot, Sylvain},
   year      = {2018},
-  publisher = {Seismological Society of America}
+  month     = aug,
+  pages     = {2687–2712}
 }
 
 @article{barbot2020frictional,
[…]
