### 2025-07-04 7881c9a
update version metadata, add installation doc
