### 2025-07-04 #8428 @shipengcheng1230
Hi @kthyng, in the [latest commit](https://github.com/shipengcheng1230/Oetqf.jl/tree/paper/paper) I have fixed the DOI missing issue. Really appreciate your feedback!
