### 2025-07-14 #8428 @willgearty
@shipengcheng1230 I think it would be useful to also include the installation details in your README file.
