### 2025-07-15 #8428 @shipengcheng1230
@willgearty thank you for the feedback, the installation details have been added, you can preview it in the paper branch ([here](https://github.com/shipengcheng1230/Oetqf.jl/tree/paper)).
