### 2025-07-15 #8428 @willgearty
:wave: @niyiyu @anowacki @ziyixi @AnantHariharan1996, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html.
