### 2025-07-16 #8597 @editorialbot
[REVIEW]: Oetqf: A Julia package for quasi-dynamic earthquake cycle simulation
**Submitting author:** @shipengcheng1230 (<a href="http://orcid.org/0000-0003-3558-458X">Pengcheng Shi</a>)
**Repository:** https://github.com/shipengcheng1230/Oetqf.jl
**Branch with paper.md** (empty if default branch): master
**Version:** 0.3.5
