### 2025-07-22 shipengcheng1230/Oetqf.jl#30 @niyiyu
JOSS review checklist
Hi @shipengcheng1230,

I am raising this issue to help you iterate your JOSS submission for this package (https://github.com/openjournals/joss-reviews/issues/8597). While I am done with most items in the checklist, there are some that may require your action.

- I feel the package may benefit from documentation improvement. Here are some points that you may consider to improve.
  - Describe the returned value for methods that applies.
  - Provide interpretations that help design the problem and understand the results, i.e., in the 2D transform fault example.
- You could provide docs about testing in your repo either maually or using GitHub Actions (see an example [here](https://kura-okubo.github.io/SeisMonitoring.jl/dev/test_actions/))
- The installation section is currently in the `paper` branch only. You may include that in your main (master) branch.

Best,
Yiyu
