+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"note: Note\nThis example corresponds to the simulations in Shi, P., Wei, M., & Barbot, S., (2022), submitted to JGR - Solid Earth. The mesh size is downgraded for speed of the document building","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"using Oetqf, SpecialFunctions, Optim","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"Generate the mesh for the transform fault, which is suited for using Okaka, (1992) equation:","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"mf = gen_mesh(Val(:RectOkada), 80e3, 8e3, 10e3, 2e3, 90.0);\nnothing #hide","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"Use Gmsh to generate the mantle mesh, which is suited for using Barbot et al., (2017) equation, with no refinement in x or y direction while cell sizes are 1.5 times progressively larger along z axes:","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"gen_gmsh_mesh(Val(:BEMHex8Mesh), -40e3, -2.5e3, -8e3, 80e3, 5e3, -22e3, 4, 3, 3;\n    output = joinpath(@__DIR__, \"mantle.vtk\"),\n    rfzh = cumprod(ones(3) * 1.5), rfy = 1.0, rfyType = \"Bump\"\n)\nma = gen_mesh(Val(:BEMHex8Mesh), joinpath(@__DIR__, \"mantle.vtk\"));\nnothing #hide","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"Compute the stress Green's function between the two meshes:","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"λ = μ = 3e10\ngffile = joinpath(@__DIR__, \"gf.h5\")\nisfile(gffile) && rm(gffile)\n@time gf₁₁ = stress_greens_function(mf, λ, μ; buffer_ratio = 1)\nh5write(gffile, \"gf₁₁\", gf₁₁) # fault -> fault\n@time gf₁₂ = stress_greens_function(mf, ma, λ, μ; buffer_ratio = 1, qtype = \"Gauss1\")\nh5write(gffile, \"gf₁₂\", gf₁₂) # fault -> mantle\n@time gf₂₁ = stress_greens_function(ma, mf, λ, μ)\nh5write(gffile, \"gf₂₁\", gf₂₁) # mantle -> fault\n@time gf₂₂ = stress_greens_function(ma, λ, μ; qtype = \"Gauss1\")\nh5write(gffile, \"gf₂₂\", gf₂₂) # mantle -> mantle","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fa
[…]
