### 2025-08-02 shipengcheng1230/Oetqf.jl#30 @shipengcheng1230
Hi @niyiyu,

Thank you so much for the great feedbacks! In the latest revision, I have improved the documentation of this package, including
- Added more public APIs
- Added description for the return values for all public APIs
- Added more description in the example notebook to better explain the problem, illustrate settings and amend various tips
- Added new pages describing testings and parameter space
- Added the installation instructions

You can view the updated documents [here](https://shipengcheng1230.github.io/Oetqf.jl/dev/). Feel free to let me know if there'a anything need further addressing and really appreciate your time and effort!
