+++ v0.3.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"note: Note\nThis example corresponds to the simulations in Shi, P., Wei, M., & Barbot, S. (2022), JGR - Solid Earth - 10.1029/2022jb024069. The mesh size is reduced to improve CI/CD speed and avoid timeouts.","category":"page"},{"location":"generated/otf-with-mantle/#Problem-statement","page":"A 2D transform fault overlaying a 3D mantle","title":"Problem statement","text":"","category":"section"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"We would like to investigate the stress interaction between a 2D plane transform fault and a 3D mantle and how it affects the seismic pattern.","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"using Oetqf, SpecialFunctions, Optim","category":"page"},{"location":"generated/otf-with-mantle/#Generate-the-meshes","page":"A 2D transform fault overlaying a 3D mantle","title":"Generate the meshes","text":"","category":"section"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"Generate the mesh for the transform fault, which is suitable for using the Okada (1992) equation. The fault is 80 km long, 8 km deep, with grid sizes of 10 km and 2 km respectively, and a dip angle of 90 degrees (vertical).","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"mf = gen_mesh(Val(:RectOkada), 80e3, 8e3, 10e3, 2e3, 90.0);\nnothing #hide","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"The mesh mf is a RectOkadaMesh, which contains the fault geometry, centroid coordinates, and other properties.","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"Use Gmsh to generate the mantle mesh, which is suitable for using the Barbot et al. (2017) equation. The volume is 80 km long, 5 km wide, and 14 km deep, with a grid size of 4 cells along the x direction, 3 cells along the y direction, and 3 cells along the z direction. There is no refinement in the x or y direction, while cell sizes are 1.5 times progressively larger along the z axis.","category":"page"},{"location":"generated/otf-with-mantle/","page":"A 2D transform fault overlaying a 3D mantle","title":"A 2D transform fault overlaying a 3D mantle","text":"gen_gmsh_mesh(Val(:BEMHex8Mesh), -40e3, -2.5e3, -8e3, 80e3, 5e3, -22e3, 4, 3, 3;\n    output = joinpath(@__DIR__, \"mantle.vtk\"),\n    rfzh = cumprod(ones(3) * 1.5), rfy = 1.0, rfyType = \"Bump\
[…]
