### 2025-08-03 shipengcheng1230/Oetqf.jl#30 @niyiyu
Hi @shipengcheng1230 ,

Thanks for your work and happy to see updates in the documentation. I will finish the review checklist and recommend for acceptance.

Yiyu
