### 2025-08-08 7c69f41
Update paper/paper.md
Co-authored-by: William Gearty <willgearty@gmail.com>
