### 2025-09-08 #8597 @willgearty
:wave: @anowacki & @jackleland, would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
