### 2025-09-09 #8597 @anowacki
Hi @willgearty.  It would be nice to get some help with our submission at https://github.com/openjournals/joss-reviews/issues/7539 as that has been stalled for a very long time.  If that is moving again I suspect I will have a bit more time to review things since I won't need to be keeping an eye on that.
