### 2025-09-25 #8597 @jackleland
Completed an initial review, shall I get back to the author with feedback (via an issue or something) or just state my findings here?
