### 2025-10-04 #8597 @willgearty
@jackleland just checking in. Will you be able to report your review findings soon?
