### 2025-10-07 shipengcheng1230/Oetqf.jl#32 @jackleland
JOSS Review Feedback/Suggestions
I've been conducting a [review](https://github.com/openjournals/joss-reviews/issues/8597#issuecomment-3327530665) as part of JOSS. Most of the way there, but some crucial problems with the code that are preventing the approval, namely that I can't get it to install. For completeness, I'll write out the outstanding points in a tick-box list and I'll be happy with the review when I've been satisfied they're addressed.

## Functionality
- [ ] *Installation:* the current installation instructions result in an error when I try to install on my dev machine
- [ ] *Functionality:* I can't execute the examples in the provided notebook due to the above point

## Documentation
- [ ] *Installation:* Installation instructions currently not working
- [ ] *Community guidelines*: There aren't, or don't appear to be, instructions on how to contribute to the software or report issues. Presumably this is via the issues page on the github repo but it would be good to clarify on the documentation itself. Similarly this info could be in a CONTRIBUTING.md in the repo itself

The error message output from trying to run the following command
``` bash
pkg> add https://github.com/shipengcheng1230/Oetqf.jl#v0.3.1
```

with a fresh environment on julia 1.11.7, is

```
PkgPrecompileError: The following 1 direct dependency failed to precompile:

Oetqf

Failed to precompile Oetqf [ba4b8404-3ddb-4226-b41e-638ed858d44e] to "/home/jleland/.julia/compiled/v1.11/Oetqf/jl_MyBTDe".
ERROR: LoadError: UndefVarError: `OrdinaryDiffEqAlgorithm` not defined in `OrdinaryDiffEq`
Stacktrace:
 [1] getproperty(x::Module, f::Symbol)
   @ Base ./Base.jl:42
 [2] top-level scope
   @ ~/.julia/packages/Oetqf/6c6Yl/src/io.jl:88
 [3] include(mod::Module, _path::String)
   @ Base ./Base.jl:562
 [4] include(x::String)
   @ Oetqf ~/.julia/packages/Oetqf/6c6Yl/src/Oetqf.jl:1
 [5] top-level scope
   @ ~/.julia/packages/Oetqf/6c6Yl/src/Oetqf.jl:28
 [6] include
   @ ./Base.jl:562 [inlined]
 [7] include_package_for_output(pkg::Base.PkgId, input::String, depot_path::Vector{String}, dl_load_path::Vector{String}, load_path::Vector{String}, concrete_deps::Vector{Pair{Base.PkgId, UInt128}}, source::Nothing)
   @ Base ./loading.jl:2881
 [8] top-level scope
   @ stdin:6
in expression starting at /home/jleland/.julia/packages/Oetqf/6c6Yl/src/io.jl:88
in expression starting at /home/jleland/.julia/packages/Oetqf/6c6Yl/src/Oetqf.jl:1
in expression starting at stdin:
```

Same error when trying to run tests directly from the repo, with either the main or paper branches.

Let me know if you'd like any more information!
