### 2025-10-08 e696856
fix: update namespace of `OrdinaryDiffEqAlgorithm`
