+++ src/io.jl
@@ -115,7 +115,7 @@ Write the solution to HDF5 file while solving the ODE. The interface
 ## Returns
 - `sol::ODESolution`: the solution object of `OrdinaryDiffEq.jl`
 """
-function wsolve(prob::ODEProblem, alg::OrdinaryDiffEq.OrdinaryDiffEqAlgorithm, file, nstep, getu, ustrs, tstr; stride::Integer=1, append::Bool=false, force::Bool=false, kwargs...)
+function wsolve(prob::ODEProblem, alg::OrdinaryDiffEq.OrdinaryDiffEqCore.OrdinaryDiffEqAlgorithm, file, nstep, getu, ustrs, tstr; stride::Integer=1, append::Bool=false, force::Bool=false, kwargs...)
     if isfile(file) && !force && !append
         @info "Overwrite existing file $(file) must set `force = true`."
         @info "Aborting computation."
