### 2025-10-08 shipengcheng1230/Oetqf.jl#32 @jackleland
Excellent, thank you for the changes. I can confirm that it now installs and I can get the tests to run; I was also able to run the  example julia script, but had to install additional dependencies (SpecialFunctions, Optim). This is obviously implied by the import statement, but **wasn't specified either in the dependency list in [compat] or the text of the example script.**

I'd also still like to see explicit contributing instructions please
