### 2025-10-09 fe9b04b
fix: version number in docs and add contributing guide
