### 2025-10-14 992846f
fix: add issue/pr template and revise contribution guide
