+++ .github/pull_request_template.md
@@ -1,0 +1,13 @@
+# Summary
+
+<!-- Provide a concise overview of the problem and how this pull request solves it. -->
+
+## Testing
+
+<!-- List the tests or commands you ran to verify the changes. Include `] test` when relevant. -->
+
+## Checklist
+
+- [ ] I updated documentation as needed.
+- [ ] I added or adjusted tests where appropriate.
+- [ ] I verified that CI passes locally or explained why it cannot be run.
