### 2025-10-15 #8597 @shipengcheng1230
Hi @willgearty,

- The authors and affiliations (including ORCIDs) are all correct.
- The release version finalized is [v0.3.5](https://github.com/shipengcheng1230/Oetqf.jl/releases/tag/v0.3.5).
- The DOI of this version is [10.5281/zenodo.17354965](https://doi.org/10.5281/zenodo.17354965), in which I edited those meta field to match the title and the author list.
- The license between the two places match (GPL 3.0).

Let me know if there is anything missing or incorrect! Thank you so much for your time!
